monet: Fix overly low chroma for tones below 90
Chroma is capped to 40 for light colors of L* 90 and above for valid
reasons. However, assigning the capped value to "chroma" fails to
account for the fact that chroma is an argument passed to the method,
so the capping effect persists across iterations. This unintentionally
affects and limits the chroma of *all* shades because the first
iteration of the loop is L*=90.

As a result, almost all colors are less chromatic than the intended
chroma >= 48. Fix the issue by assigning capped chroma to a temporary
per-iteration variable.

Change-Id: I18102d0072c89535e27f23fdb98af7c83864ff7e
Signed-off-by: Dmitrii <bankersenator@gmail.com>
Signed-off-by: Romeo <rbesar210311@gmail.com>